Large modification
Update cards.png to set card images to 40x59. This works for current size,
but cards need to be larger and image quality is currently in the shitter.

Update global.js to create a div with correct classes to allow me to key off
of the class names set up for the new sprite. Again, this is going to be
fucked as soon as we change card sizes.

Update style.css (i need to change this stupid ass name) for background
positions. Unfortunately, firefox is too cool to use
background-position-x and background-position-y and theres no way in hell
we're going to set background positions for each and every card, so we need
to make a js function to handle this without writing a million lines of shit
css.

Minor updates to tripeaks.js to support the above.